[refactor] UserIdAndRole 확장
문제 사항 :
 - comment 도메인에서 사용했던 UserIdAndRole를 확장할 필요성 (다른 도메인에서도 사용할 수 있도록)

변경 사항 :
 - UserIdAndRole common 패키지로 이동
 - 확장을 위핸 AuthenticationService 생성 (member인지 manager인지 구분, role 관련 없이 officeid 뽑아냄)
